Create #33 Parity_Control
The ASCII numbers provided on the site exceed the normal english ASCII range. I assume it is because the creator of the problem does not use the same alphabet. So instead of importing the whole ASCII range, because i need only the few letters, I only put those few in the Decypher() function.